HardwareSerial: A new class derived from class Serial
Because the NMEA0183 component being developed for Arduino
depends on a HardwareSerial class it has been necessary
to implement a HardwareSerial with exactly the features
needed.

For now it is more or less a thin wrapper of the SerialDev
component - directly accessing the USART 1 pins.

It could easily be changed to support more instances (a.k.a.
USARTS)

Notice how it supports a blocked read.