spiDev: A FreeRTOS compliant SPI driver for stm32f10x controllers.
The driver provides an init function spiInit().
Despite the fact a SPI device is given, the driver currently only supports
SPI2.

The function spiReadWrite() is the function to use for transmit and/or
reception of SPI data. The function won't return until all data have been
transmitted/received. The function must NOT be called from within interrupt
context as it might block.

The actual SPI communication is handled by an IRQ handler.