Arduino.h: A garbage file making some Arduino libraries live outside Arduino
In the Arduino world include of Arduino.h ensures a lot of other .h files
are included.

Some of these are stub'ed here by adding some of the necessary defines.
Some are included to avoid changing the libraries.

In general only Arduino specific includes are made. When a library use e.g.
memcmp string.h is not added to Arduino.h but added directly into the
library needing it.