nmea0183Task: A FreeRTOS task responsible for reception of NMEA0183 packages
The task creates an nmea0183 object (known from Arduino world) which will
read data from the provided UART. When a complete message is received it
is given to the message handler which takes over from here.

The nmea0183 object is told to use blocking read. I.e. when no data is
available the nmea0183Task instance is not using CPU resources