mcpNMEA2000: A FreeRTOS/STM32F10x variant of NMEA2000_mcp made for Arduino.
mcpNMEA2000 is a realization of the virtual class tNMEA2000.

It has been implemented as a singleton. Use getInstance() to access it.

All can frames (both incoming and outgoing) are exchanged between the logic
 in class tNMEA2000 and a special FreeRTOS task.
I.e. when the logic want to send some frames they are just added to the out
queue. The FreeRTOS task being responsible for communication to the MCP2515
device will pick out frames from the queue when ever the MCP2515 is ready.

In order to be in control of the order of frames send/received only on
RX buffer and one TX buffer is used in MCP2515.

The module use the interrupt feature of the MCP2515. Currently the ISR pin
being used is hardcoded to pin B7.