nmea0183Handler: HandleVDM(), fixed handling of undefined Heading in Pos Reports.
In Pos Reports (Msg type 1,2,3 and 18) the Heading may have the value 511
indicating undefined.

Converting it uncritical to Rad's is not the right approach when the
receiver is a B&G plotter.

Think it is better then to just insert COG into HDG field too.
We won't detect a ship sailing backwards, but that's not different from
a ship not sending HDG